+++ ignite/services/plugin/internal/testdata/execute_fail/main.go
@@ -1,0 +1,44 @@
+package main
+
+import (
+	"context"
+	"errors"
+
+	hplugin "github.com/hashicorp/go-plugin"
+
+	"github.com/ignite/cli/v28/ignite/services/plugin"
+)
+
+type app struct{}
+
+func (app) Manifest(ctx context.Context) (*plugin.Manifest, error) {
+	return &plugin.Manifest{
+		Name: "execute_fail",
+	}, nil
+}
+
+func (app) Execute(ctx context.Context, cmd *plugin.ExecutedCommand, api plugin.ClientAPI) error {
+	return errors.New("fail")
+}
+
+func (app) ExecuteHookPre(ctx context.Context, h *plugin.ExecutedHook, api plugin.ClientAPI) error {
+	return nil
+}
+
+func (app) ExecuteHookPost(ctx context.Context, h *plugin.ExecutedHook, api plugin.ClientAPI) error {
+	return nil
+}
+
+func (app) ExecuteHookCleanUp(ctx context.Context, h *plugin.ExecutedHook, api plugin.ClientAPI) error {
+	return nil
+}
+
+func main() {
+	hplugin.Serve(&hplugin.ServeConfig{
+		HandshakeConfig: plugin.HandshakeConfig(),
+		Plugins: map[string]hplugin.Plugin{
+			"execute_fail": plugin.NewGRPC(&app{}),
+		},
+		GRPCServer: hplugin.DefaultGRPCServer,
+	})
+}